requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pydantic
python-dateutil
numpy
pandas